refactor: renderPptx.ts をノードタイプ別に分割
- renderPptx.ts を 1,171行から 110行に削減
- nodes/ 配下にノードタイプごとのレンダラーを分離
  - text, image, box, table, shape, chart
  - timeline, matrix, tree, flow, processArrow
- utils/ 配下に共通ユーティリティを抽出
  - backgroundBorder: 背景・ボーダー描画
  - shapeDrawing: 円・線描画
  - textDrawing: テキスト描画
- types.ts に RenderContext 型を定義
