+++ src/renderPptx/nodes/box.ts
@@ -1,0 +1,2 @@
+// box, vstack, hstack ノードは背景描画のみで、子要素は renderPptx.ts の renderNode() で再帰処理
+// このファイルは将来の拡張用に作成するが、現在は特別な処理を必要としない
