+++ src/renderPptx/nodes/index.ts
@@ -1,0 +1,10 @@
+export { renderTextNode } from "./text";
+export { renderImageNode } from "./image";
+export { renderTableNode } from "./table";
+export { renderShapeNode } from "./shape";
+export { renderChartNode } from "./chart";
+export { renderTimelineNode } from "./timeline";
+export { renderMatrixNode } from "./matrix";
+export { renderTreeNode } from "./tree";
+export { renderFlowNode } from "./flow";
+export { renderProcessArrowNode } from "./processArrow";
