+++ src/renderPptx/utils/index.ts
@@ -1,0 +1,5 @@
+export { renderBackgroundAndBorder } from "./backgroundBorder";
+export { drawCircle, drawLine } from "./shapeDrawing";
+export type { CircleOptions, LineOptions } from "./shapeDrawing";
+export { drawSimpleText } from "./textDrawing";
+export type { SimpleTextOptions } from "./textDrawing";
